refactored parsing and validation code in nostr/types.go to separate nostr/parser.go file
